requirements.txt
pandas
numpy
matplotlib
scikit-learn
openai
python-dotenv
chromadb